Add files via upload
Project Streamlit is a YouTube video downloader app that allows users to download YouTube videos by providing a URL. Main functions include:

YouTube video download: Users can enter the YouTube video URL and display the screen where the video will be saved. If the directory does not exist or is not provided, the app will make the user's desktop. . The workbench deletes the previous video and resets the access point, allowing the user to start a new download. Streamlit's user-friendly interface.